Changed Database to PostgresSQL and implemented JWT for secure authentication

+++ employee_management/employee/models.py
@@ -13,7 +13,7 @@ class Employee(models.Model):
     first_name = models.CharField(max_length=100,null=False,blank=False)
     last_name = models.CharField(max_length=100,null=False,blank=False)
     email = models.EmailField(max_length=100,null=False,blank=False)
-    phone = models.IntegerField(null=False,blank=False)
+    phone = models.BigIntegerField(null=False,blank=False)
     address = models.CharField(max_length=100,null=False,blank=False)
     department = models.ForeignKey(Department,on_delete=models.CASCADE)
     role = models.ForeignKey(Role,on_delete=models.CASCADE)
